requirements.txt
python-novaclient
keystoneauth1
fastapi